refactor(rtp_recv): split main_rtp_recv.cpp by concern
Splits the ~1735-line main_rtp_recv.cpp into concern-sized files:

  cli.{hpp,cpp}                     — CliOptions struct, parse_cli,
                                      print_usage
  decode_helpers.{hpp,cpp}          — shared decode functions used by
                                      both pipelines: coefficient
                                      selection, color pipeline setup,
                                      decode_to_rgb/planar, frame dump
  pipeline_single_threaded.{hpp,cpp} — v1 single-threaded main loop
  pipeline_multi_threaded.{hpp,cpp}  — v2 ReceiverState + recv/decode
                                      threads + threaded main loop
  smoke_tests.{hpp,cpp}             — run_smoke_tests + all 5 tests
  main_rtp_recv.cpp                 — 39-line entry point + dispatcher

No behavior change. 582/582 ctest pass. Profile sweep at threads=
{2,4,6} shows no regression (deltas within noise — the refactoring
touches only app-layer file organization, not the core decode path).

Closes #203.
