+++ source/apps/rtp_recv/CMakeLists.txt
@@ -2,6 +2,11 @@ cmake_policy(SET CMP0076 NEW)
 target_sources(open_htj2k_rtp_recv
     PRIVATE
         main_rtp_recv.cpp
+        cli.cpp
+        decode_helpers.cpp
+        smoke_tests.cpp
+        pipeline_single_threaded.cpp
+        pipeline_multi_threaded.cpp
         rtp_socket.cpp
         rfc9828_parser.cpp
         frame_handler.cpp
