+++ source/apps/rtp_recv/smoke_tests.hpp
@@ -1,0 +1,14 @@
+// Copyright (c) 2026, Osamu Watanabe
+// All rights reserved.
+//
+// Licensed under the same BSD 3-Clause terms as the rest of OpenHTJ2K.
+
+#pragma once
+
+#include "cli.hpp"
+
+namespace open_htj2k::rtp_recv {
+
+int run_smoke_tests(const CliOptions& opts);
+
+}  // namespace open_htj2k::rtp_recv
